backend/pyproject.toml
[project]
name = "autoclipeval-backend"
version = "0.1.0"
description = "Backend for AutoClipEval, based on FastAPI Full Stack Template."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.82.1",
    "pydantic[bson]>=2.11.5",
    "pymongo[srv]>=4.13.0",
    "python-dotenv>=1.1.0",
    "smolagents[all,mcp,openai,toolkit]>=1.17.0",
    "uvicorn[standard]>=0.29.0",
    "yt-dlp>=2025.5.22",
] 

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.12",
]
